Separates ShortcutConfigurableNonRestrictiveTests
